script_v2/requirements.txt
google-genai>=1.0.0
requests>=2.31.0
mutagen>=1.47.0
Pillow>=10.0.0
python-dotenv>=1.0.0
